front: refactor RejectApplication modal to have correct modal buttons
